requirements.txt
# Core deep learning frameworks
tensorflow>=2.13.0
keras>=2.13.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data manipulation and visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Web framework for API
flask>=3.0.0
flask-cors>=4.0.0

# Additional utilities
tqdm>=4.65.0
h5py>=3.9.0
